+++ cmu_graphics/meta/updates.json
@@ -1,1 +1,1 @@
-{"last_attempt": 1733022779.475944, "most_recent_version": "1.1.37"}
+{"last_attempt": 1733297863.485016, "most_recent_version": "1.1.38"}
