+++ tp06.py
@@ -1,6 +1,6 @@
 from rembg import remove
 from PIL import Image
-#from cmu_graphics import CMUImage
+from cmu_graphics import CMUImage
 import os
 import cv2
 import numpy as np
@@ -13,15 +13,19 @@ def loadImage(inputPath):
     i = 0
     for item in os.listdir(folder_path):
         #
-        img_path = os.path.join(folder_path, item)
-        input_image = Image.open(img_path)
+        
         #print("name: ", img_path)
-        print("loadimage item: ", item)
-        output_image = remove(input_image)
+        
         #output_folder = f"C:\\Users\\irisy\\Desktop\\15-112\\termProject\\Term-Project\\removedbgClothing\\{i}"
-        output_folder =  f"C:\\Users\\irisy\\Desktop\\15-112\\termProject\\term-project-v2\\remobedtest\\{i}"
+        output_folder =  f"C:\\Users\\irisy\\Desktop\\15-112\\termProject\\term-project-v2\\processed\\"
         ##print('heloooooooooo', item)
         output_path = output_folder+item[:-4] +".png"
+        if os.path.exists(output_path):
+            continue
+        img_path = os.path.join(folder_path, item)
+        input_image = Image.open(img_path)
+        print("loadimage item: ", item)
+        output_image = remove(input_image)
         output_image.save(output_path)
         result.append(output_path)
         print("result: ", result)
